_dispatcher module and type hints
